Increase performance for molecules with many atoms
Finding for rotational axes between atoms is an expensive procedure for
large molecules, as it first has to iterate over O(n^2) pairs of atoms
and then check whether the rotations exist (at least O(n)).

These changes speed the process up in a few ways: first, we skip
checking rotations if the atoms in the pair are far apart, and second,
we skip checking higher-order rotations if a C₂ rotation isn't found.
In testing, this seems to lead to a ~5× performance improvement for
buckminsterfullerene (C₆₀, an extreme case).